requirements.txt
pyautogui==0.9.54 
pyyaml==6.0.2 
Pillow==11.2.1 
img2pdf==0.6.0 
pymsgbox==1.0.9 
pytweening==1.2.0 
pyscreeze==1.0.1 
pygetwindow==0.0.9 
mouseinfo==0.1.3 
pikepdf==9.7.0
pyrect==0.2.0 
pyperclip==1.9.0 
Deprecated==1.2.18 
lxml==5.3.2 
packaging==24.2 
wrapt==1.17.2